Committing svn revision 37511: Updating 7.1.0 tree to 7.1.1 beta1, merging 36420:37267 ZCP-7269

+++ provider/libserver/ECDatabaseUpdate.cpp
@@ -1291,8 +1291,8 @@ ECRESULT UpdateDatabaseCreateABChangesTable(ECDatabase *lpDatabase)
 	er = lpDatabase->DoInsert(Z_TABLEDEF_ABCHANGES);
 	if (er != erSuccess)
 		goto exit;
-
-	strQuery = "SELECT id, sourcekey, parentsourcekey, change_type, moved_from FROM changes WHERE change_type & " + stringify(ICS_AB) + " AND parentsourcekey=0x00000000AC21A95040D3EE48B319FBA75330442500000000040000000100000000000000";
+		
+	strQuery = "SELECT id, sourcekey, parentsourcekey, change_type FROM changes WHERE change_type & " + stringify(ICS_AB) + " AND parentsourcekey=0x00000000AC21A95040D3EE48B319FBA75330442500000000040000000100000000000000";
 	er = lpDatabase->DoSelect(strQuery, &lpResult);
 	if (er != erSuccess)
 		goto exit;
@@ -1311,14 +1311,10 @@ ECRESULT UpdateDatabaseCreateABChangesTable(ECDatabase *lpDatabase)
 		syncIds.push_back(ulId);
 
 		strQuery = "INSERT INTO abchanges (id, sourcekey, parentsourcekey, change_type";
-		if (lpDBRow[4] != NULL && lpDBLen[4] != 0)
-			strQuery += ", moved_from";
 		strQuery += (string)") VALUES (" + lpDBRow[0] + ", " +
 								   lpDatabase->EscapeBinary((unsigned char*)lpDBRow[1], lpDBLen[1]) + ", " +
 								   lpDatabase->EscapeBinary((unsigned char*)lpDBRow[2], lpDBLen[2]) + ", " +
 								   lpDBRow[3];
-		if (lpDBRow[4] != NULL && lpDBLen[4] != 0)
-			strQuery += ", " + lpDatabase->EscapeBinary((unsigned char*)lpDBRow[4], lpDBLen[4]);
 		strQuery += ")";
         queries.push_back(strQuery);
 	}
@@ -1805,10 +1801,12 @@ exit:
 ECRESULT UpdateDatabaseSyncTimeIndex(ECDatabase *lpDatabase)
 {
 	ECRESULT er = erSuccess;
+	bool bHaveIndex;
 
-	er = lpDatabase->DoUpdate("ALTER TABLE syncs ADD INDEX sync_time (`sync_time`);");
-	if (er != erSuccess)
-		goto exit;
+	// There are upgrade paths where the sync_time key already exists.
+	er = lpDatabase->CheckExistIndex("syncs", "sync_time", &bHaveIndex);
+	if (er == erSuccess && !bHaveIndex)
+		er = lpDatabase->DoUpdate("ALTER TABLE syncs ADD INDEX sync_time (`sync_time`)");
 
 exit:
 	return er;
@@ -1818,8 +1816,12 @@ exit:
 ECRESULT UpdateDatabaseAddStateKey(ECDatabase *lpDatabase)
 {
 	ECRESULT er = erSuccess;
+	bool bHaveIndex;
 
-	er = lpDatabase->DoUpdate("ALTER TABLE changes ADD UNIQUE KEY `state` (`parentsourcekey`,`id`)");
+	// There are upgrade paths where the state key already exists.
+	er = lpDatabase->CheckExistIndex("changes", "state", &bHaveIndex);
+	if (er == erSuccess && !bHaveIndex)
+		er = lpDatabase->DoUpdate("ALTER TABLE changes ADD UNIQUE KEY `state` (`parentsourcekey`,`id`)");
 
 	return er;
 }
@@ -2316,9 +2318,15 @@ ECRESULT UpdateDatabaseConvertChanges(ECDatabase *lpDatabase)
 {
 	ECRESULT er = erSuccess;
 	std::string strQuery;
-
-	strQuery = "ALTER TABLE changes DROP COLUMN moved_from, DROP key moved";
-	er = lpDatabase->DoDelete(strQuery);
+	bool bDropColumn;
+	
+	// In some upgrade paths the moved_from column doesn't exist. We'll
+	// check so no error (which we could ignore) will be logged.
+	er = lpDatabase->CheckExistColumn("changes", "moved_from", &bDropColumn);
+	if (er == erSuccess && bDropColumn) {
+		strQuery = "ALTER TABLE changes DROP COLUMN moved_from, DROP key moved";
+		er = lpDatabase->DoDelete(strQuery);
+	}
 
 	return er;
 }
